agents/dok2u-agent/requirements.txt
fastapi
uvicorn
openai
python-multipart
pydantic
ffmpeg-python
pydub
tqdm
pillow
pytesseract
whisper
python-dotenv
jinja2
chromadb
python-docx
instagrapi
google-generativeai
google-cloud-aiplatform
google-cloud-storage
PyPDF2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
moviepy